Rename `device.py` to `devices.py`
