+++ src/metatensor/models/utils/devices.py
